feat: Creation rectangle has keypoints size
